KJHH-1246: Refactoring move KutsuminenPage to kutsuminen-folder
